Renamed tests to follow the same convention
