+++ src/main/java/com/codurance/allaboard/core/model/catalogue/LearningPath.java
@@ -3,7 +3,6 @@ package com.codurance.allaboard.core.model.catalogue;
 import com.codurance.allaboard.core.model.topic.Topic;
 import java.io.Serializable;
 import java.util.List;
-import java.util.Set;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
