+++ src/test/java/com/codurance/allaboard/acceptance/utils/WebAcceptanceTestTemplateShould.java
@@ -1,4 +1,4 @@
-package com.codurance.allaboard.e2e.utils;
+package com.codurance.allaboard.acceptance.utils;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -7,13 +7,13 @@ import org.springframework.test.util.ReflectionTestUtils;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
-public class WebAcceptanceE2ETestTemplateShould {
+public class WebAcceptanceTestTemplateShould {
 
-    WebAcceptanceE2ETestTemplate utils;
+    WebAcceptanceTestTemplate utils;
 
     @BeforeEach
     private void setUp() {
-        this.utils = new WebAcceptanceE2ETestTemplate();
+        this.utils = new WebAcceptanceTestTemplate();
         ReflectionTestUtils.setField(utils, "apiUrlPrefix", "apiVersion");
     }
