+++ src/test/java/com/codurance/allaboard/core/unit/SaveSurveyShould.java
@@ -7,7 +7,6 @@ import static org.mockito.Mockito.verify;
 import com.codurance.allaboard.core.actions.survey.SaveSurvey;
 import com.codurance.allaboard.core.model.survey.Survey;
 import com.codurance.allaboard.core.model.survey.Surveys;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
