+++ src/test/java/com/codurance/allaboard/web/acceptance/TopicDetailsFetchFeature.java
@@ -1,6 +1,6 @@
 package com.codurance.allaboard.web.acceptance;
 
-import com.codurance.allaboard.e2e.utils.WebAcceptanceE2ETestTemplate;
+import com.codurance.allaboard.acceptance.utils.WebAcceptanceTestTemplate;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
@@ -19,7 +19,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class TopicDetailsFetchingFeature extends WebAcceptanceE2ETestTemplate {
+public class TopicDetailsFetchFeature extends WebAcceptanceTestTemplate {
 
     @LocalServerPort
     private int port;
